# [Performance] don't check ancestry if block is in the current chain